refactor: move bn254 public-input hashing to the backend behind a reusable trait
Public-input instance hashing (HashConfig::hash_field_elements + the
Skyscraper/SHA/Keccak/BLAKE3/Poseidon2 bodies + the DST tag) was bn254-welded
code sitting in common. Move it to provekit-backend-bn254 (field_hash.rs), where
Bn254Field::hash_public_inputs already wants it, and add a generic
PublicInputsHash trait in common so any backend gets ergonomic
pi.hash_bytes::<Field>(config) for free — the field-selecting type parameter is
on the method (not the trait), so the impl stays a plain, coherence-clean
impl<F: Field> for PublicInputs<F> and the call site names the field once. The
byte-exact KAT regression suite moves with the hashing. common drops its sha2 and
blake3 deps; bn254 stays byte-identical.
