+++ provekit/backend/bn254/Cargo.toml
@@ -27,13 +27,18 @@ noirc_abi.workspace = true
 ark-bn254.workspace = true
 ark-ff.workspace = true
 ark-std.workspace = true
+poseidon2.workspace = true
+skyscraper.workspace = true
 whir.workspace = true
 
 # 3rd party
 anyhow.workspace = true
+blake3.workspace = true
 postcard.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+sha2.workspace = true
+sha3.workspace = true
 tracing.workspace = true
 
 # Target-specific dependencies: only on non-WASM targets
@@ -43,6 +48,9 @@ nargo = { workspace = true, optional = true }
 noir_artifact_cli = { workspace = true, optional = true }
 mavros-vm.workspace = true
 mavros-artifacts.workspace = true
+
+[dev-dependencies]
+proptest.workspace = true
 
 [lints]
 workspace = true
