+++ provekit/backend/bn254/src/lib.rs
@@ -4,6 +4,7 @@
 //! `solver` (witness solving), and the `prove`/`verify` orchestration.
 
 mod field;
+mod field_hash;
 mod frontend;
 #[cfg(not(target_arch = "wasm32"))]
 mod mavros_prove;
