+++ provekit/common/Cargo.toml
@@ -41,8 +41,6 @@ rayon.workspace = true
 ruint.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-blake3.workspace = true
-sha2.workspace = true
 sha3.workspace = true
 tracing.workspace = true
 zerocopy.workspace = true
